Move file from http_wrapper -> http_request
