+++ shoes-core/lib/shoes/http_request.rb
